Update challenge to Promise
